Add Partial Views

+++ MyLeasing.Web/Controllers/OwnersController.cs
@@ -138,7 +138,7 @@ namespace MyLeasing.Web.Controllers
                 return NotFound();
             }
 
-            var view = new EditUserViewModel
+            var model = new EditUserViewModel
             {
                 Address = owner.User.Address,
                 Document = owner.User.Document,
@@ -148,30 +148,30 @@ namespace MyLeasing.Web.Controllers
                 PhoneNumber = owner.User.PhoneNumber
             };
 
-            return View(view);
+            return View(model);
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(EditUserViewModel view)
+        public async Task<IActionResult> Edit(EditUserViewModel model)
         {
             if (ModelState.IsValid)
             {
                 var owner = await _dataContext.Owners
                     .Include(o => o.User)
-                    .FirstOrDefaultAsync(o => o.Id == view.Id);
+                    .FirstOrDefaultAsync(o => o.Id == model.Id);
 
-                owner.User.Document = view.Document;
-                owner.User.FirstName = view.FirstName;
-                owner.User.LastName = view.LastName;
-                owner.User.Address = view.Address;
-                owner.User.PhoneNumber = view.PhoneNumber;
+                owner.User.Document = model.Document;
+                owner.User.FirstName = model.FirstName;
+                owner.User.LastName = model.LastName;
+                owner.User.Address = model.Address;
+                owner.User.PhoneNumber = model.PhoneNumber;
 
                 await _userHelper.UpdateUserAsync(owner.User);
                 return RedirectToAction(nameof(Index));
             }
 
-            return View(view);
+            return View(model);
         }
 
 
@@ -186,44 +186,22 @@ namespace MyLeasing.Web.Controllers
 
             var owner = await _dataContext.Owners
                 .Include(o => o.User)
+                .Include(o => o.Properties)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (owner == null)
             {
                 return NotFound();
             }
 
+            if (owner.Properties.Count != 0)
+            {
+                ModelState.AddModelError(string.Empty, "Owner cam´t be delele because it has properties.");
+                return RedirectToAction("Index");
+            }
+
             _dataContext.Owners.Remove(owner);
             await _dataContext.SaveChangesAsync();
             await _userHelper.DeleteUserAsync(owner.User.Email);
-            return RedirectToAction(nameof(Index));
-        }
-
-
-        //public async Task<IActionResult> Delete(int? id)
-        //{
-        //    if (id == null)
-        //    {
-        //        return NotFound();
-        //    }
-
-        //    var owner = await _dataContext.Owners
-        //        .FirstOrDefaultAsync(m => m.Id == id);
-        //    if (owner == null)
-        //    {
-        //        return NotFound();
-        //    }
-
-        //    return View(owner);
-        //}
-
-        // POST: Owners/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var owner = await _dataContext.Owners.FindAsync(id);
-            _dataContext.Owners.Remove(owner);
-            await _dataContext.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
@@ -264,6 +242,7 @@ namespace MyLeasing.Web.Controllers
 
 
             }
+            model.PropertyTypes = _combosHelper.GetComboPropertyTypes();
             return View(model);
      
         }
@@ -424,6 +403,8 @@ namespace MyLeasing.Web.Controllers
 
             }
 
+            model.Lessees = _combosHelper.GetComboLessees();
+
             return View(model);
         }
 
@@ -515,12 +496,19 @@ namespace MyLeasing.Web.Controllers
 
             var property = await _dataContext.Properties
                 .Include(p => p.Owner)
+                .Include(p => p.PropertyImages)
+                .Include(p => p.Contracts)
                 .FirstOrDefaultAsync(pi => pi.Id == id.Value);
             if (property == null)
             {
                 return NotFound();
             }
-
+            if (property.Contracts.Count != 0)
+            {
+                ModelState.AddModelError(string.Empty, "The property can´t be deleted because it has contracts.");
+                                return RedirectToAction("Details", "Owners", new { id = property.Owner.Id });
+            }
+            _dataContext.PropertyImages.RemoveRange(property.PropertyImages);
             _dataContext.Properties.Remove(property);
             await _dataContext.SaveChangesAsync();
             //return RedirectToAction($"{nameof(Details)}/{property.Owner.Id}");
